requirements.txt

# Discord bot dependencies
discord.py==2.5.2
flask==3.0.2
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
typing-extensions==4.13.0
yarl==1.15.2
psycopg2-binary==2.9.6
gTTS==2.5.4
PyNaCl==1.5.0
emoji==2.2.0
pydub==0.25.1
openai==1.43.0
httpx==0.27.2

# LoRA training and model serving dependencies
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
datasets>=2.14.0
evaluate>=0.4.0
scikit-learn>=1.3.0
tensorboard>=2.14.0
tqdm>=4.65.0
sentencepiece>=0.1.99
safetensors>=0.3.0
huggingface-hub>=0.17.0

# Additional utilities
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0